requirements.txt
alembic==1.12.0
blinker==1.6.3
blitz-bayesian-pytorch==0.2.8
cdt==0.6.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cloudpickle==2.2.1
cmake==3.27.7
contourpy==1.1.1
cycler==0.12.1
databricks-cli==0.18.0
docker==6.1.3
entrypoints==0.4
filelock==3.12.4
Flask==2.3.3
fonttools==4.43.1
gitdb==4.0.10
GitPython==3.1.38
GPUtil==1.4.0
greenlet==3.0.0
gunicorn==20.1.0
higher==0.2.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
jax==0.4.11
jaxlib==0.4.18
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lit==17.0.2
Mako==1.2.4
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.0
ml-dtypes==0.3.1
mlflow==2.3.2
mpmath==1.3.0
multipledispatch==1.0.0
networkx==3.1
numpy==1.24.4
numpyro==0.12.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.1
patsy==0.5.3
Pillow==10.1.0
protobuf==4.24.4
pyarrow==11.0.0
pyemd==1.0.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
querystring-parser==1.2.4
requests==2.31.0
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
skrebate==0.62
smmap==5.0.1
SQLAlchemy==2.0.22
sqlparse==0.4.4
statsmodels==0.14.0
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.2.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
websocket-client==1.6.4
Werkzeug==3.0.0
zipp==3.17.0
